fix(Asciidoctor-panel): inconsistent module layout.  this may be the cause of the 404 on the asciidoctor-panel html template. Hopefully.
