+++ src/lib/asciidoctor-panel/asciidoctor-panel.component.scss
