+++ src/lib/asciidoctor-panel/asciidoctor-panel.module.ts
@@ -1,0 +1,12 @@
+import 'asciidoctorjs-web-repack/asciidoctor-all.min'
+import {NgModule} from "@angular/core";
+import {AsciidoctorComponent} from "./asciidoctor-panel.component";
+
+
+@NgModule({
+  declarations: [AsciidoctorComponent],
+  exports: [AsciidoctorComponent],
+})
+export class AsciidoctorPanelModule {
+
+}
