+++ src/lib/asciidoctor-panel/index.ts
@@ -1,1 +1,3 @@
-export * from './asciidoctor-panel'
+export * from './asciidoctor-panel.component'
+export * from './asciidoctor-renderer'
+export * from './asciidoctor-panel.module'
